Renamed bitPosLro() function to bit_position_lro() and updated code.

+++ lib/bitops.h
@@ -8,7 +8,7 @@
 int word32_bit(uint32_t word, size_t n);
 int word64_bit(uint64_t word, size_t n);
 ssize_t bit_position_rro(size_t wsize, size_t n, size_t rot);
-int bit_position_lro(size_t wsize, size_t n, size_t rot);
+ssize_t bit_position_lro(size_t wsize, size_t n, size_t rot);
 
 bool * next_combination(bool *comb, size_t size);
